signal: Protect tsk->sighand with {get,put}_sighand()
This patch enables us to prevent tsk->sighand being changed without
requiring that we acquire the shared siglock. The motivation for this
change is a later patch where we try to avoid taking the shared
siglock at all because it's a major point of contention in
multithreaded apps.

Currently, in order to safely operate on a target task's signal
structures we must acquire the shared siglock, tsk->sighand->siglock.
However, there is potential for someone to modify tsk's ->sighand
pointer while we're derefencing it; this can happen in two places,

	1. __exit_signal() - when a task is exiting
	2. de_thread() - when a task exec()'s

At the moment we handle this in __lock_task_sighand(),

        rcu_read_lock();
        for (;;) {
                sighand = rcu_dereference(tsk->sighand);
                if (unlikely(sighand == NULL))
                        break;

                spin_lock_irqsave(&sighand->siglock, *flags);
                if (likely(sighand == tsk->sighand))
                        break;
                spin_unlock_irqrestore(&sighand->siglock, *flags);
        }
        rcu_read_unlock();

which ensures that we actually lock the correct spinlock if someone
changes tsk->sighand under us. The problem is that this scheme
requires us to grab the shared siglock, and as we're trying to move
away from acquiring the shared siglock wherever possible, this isn't
the best approach.

This patch introduces a new per-thread lock that protects a task's
sighand pointer. As the lock is per-thread it should scale much better
than using the shared siglock. Note that the lock does not prevent the
struct_sighand that tsk->sighand points to from being modified,
rather, it prevents tsk->sighand from being modified therefore making
it safe to dereference it. The shared siglock still needs to be
acquired when synchronization is needed for *tsk->sighand and
*tsk->signal.

get_sighand() and put_sighand() are two new helper functions for use
when dereferencing tsk->sighand that hide the detais of the new lock.

The new idiom for grabbing the shared siglock is thus,

	sighand = get_sighand(tsk);
	if (!sighand)
		return -ESRCH;

	spin_lock_irqsave(&sighand->siglock, flags);

	<... do stuff under shared siglock ...>

	spin_unlock_irqrestore(&sighand->siglock, flags);
	put_sighand(tsk);

Signed-off-by: Matt Fleming <matt.fleming@linux.intel.com>